StatId Attribut Long id -> Integer id

+++ src/main/java/htwberlin/webtech/persistence/StatRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 @Repository
 public interface StatRepository extends CrudRepository<Stat, StatId> {
 
-    Iterable<Stat> findById_Id(Long id);
+    Iterable<Stat> findById_Id(Integer id);
     Iterable<Stat> findById_Rating(Integer rating);
 
 }
